+++ packages/prisma/test/__fixture__/app/app.module.ts
@@ -6,6 +6,7 @@ import { UsersCursorDefaultSortController } from './users-cursor-default-sort.co
 import { UsersCursorMultiSortController } from './users-cursor-multi-sort.controller';
 import { UsersCursorNoLimitController } from './users-cursor-no-limit.controller';
 import { UsersController } from './users.controller';
+import { UsersRouteDefaultsController, UsersRouteDefaultsService } from './users-route-defaults.controller';
 import { UsersService, PRISMA_CLIENT } from './users.service';
 
 @Module({})
@@ -35,8 +36,9 @@ export class AppModule {
         UsersCursorNoLimitController,
         UsersCursorDefaultSortController,
         UsersCursorMultiSortController,
+        UsersRouteDefaultsController,
       ],
-      providers: [{ provide: PRISMA_CLIENT, useValue: prismaClient }, UsersService],
+      providers: [{ provide: PRISMA_CLIENT, useValue: prismaClient }, UsersService, UsersRouteDefaultsService],
       exports: [UsersService],
     };
   }